added favorite migration

+++ database/migrations/2024_07_08_111059_add_user_id_to_recipes_table.php
@@ -3,37 +3,30 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
-use Illuminate\Support\Facades\DB;
 
 class AddUserIdToRecipesTable extends Migration
 {
     /**
      * Run the migrations.
-     *
-     * @return void
      */
-    public function up()
+    public function up(): void
     {
         Schema::table('recipes', function (Blueprint $table) {
-            // Add user_id column
-            $table->unsignedBigInteger('user_id')->after('id');
-
-            // Optionally, add a foreign key constraint
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+          
+            if (!Schema::hasColumn('recipes', 'user_id')) {
+                $table->unsignedBigInteger('user_id')->after('id');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            }
         });
     }
 
     /**
      * Reverse the migrations.
-     *
-     * @return void
      */
-    public function down()
+    public function down(): void
     {
         Schema::table('recipes', function (Blueprint $table) {
-     
             $table->dropForeign(['user_id']);
-    
             $table->dropColumn('user_id');
         });
     }
